==> src/05_github-networks/02_international-networks/03_country-networks/00_antiquated/04_country-networks-cumulative-old.Rmd <==
---
title: "Create Edgelists"
output: html_document
---

## Loading Data 

We used this code to pull data for the construction of Github network data (cumulatively.

```{r loading edgelist data, warning=FALSE, echo=FALSE}
rm(list = ls())

# load packages 
for (pkg in c("tidyverse", "igraph", "visNetwork", "bc3net", "data.table", "R.utils", "RPostgreSQL",
              "cowplot", "maditr", "lubridate", "linkprediction", "poweRlaw")) {library(pkg, character.only = TRUE)}

# connect to postgresql to get data (in rivanna)
conn <- dbConnect(drv = PostgreSQL(), dbname = "sdad_data", 
                  host = "sdad.policy-analytics.net", port = 5436, 
                  user = Sys.getenv("db_userid"), password = Sys.getenv("db_pwd"))

# query the bipartite edgelist data from github data  
bp_edgelist <- dbGetQuery(conn, "SELECT login, slug, year, weight 
                                 FROM github.intl_dyn_bp_el_new 
                                 WHERE year = 2008")
                                #WHERE year > 2007 AND year < 2019")

# disconnect from postgresql
dbDisconnect(conn)

# this code is taken from network scientists to reduce bipartite networks to single mode networks 
bp_matrix <- spMatrix(nrow=length(unique(bp_edgelist$login)),
              ncol=length(unique(bp_edgelist$slug)),
              i = as.numeric(factor(bp_edgelist$login)),
              j = as.numeric(factor(bp_edgelist$slug)),
              x = rep(1, length(as.numeric(bp_edgelist$login))))
row.names(bp_matrix) <- levels(factor(bp_edgelist$login))
colnames(bp_matrix) <- levels(factor(bp_edgelist$slug))

# we then multiply these incidence matrices and impute zeroes to prevent self loops 
login_matrix <- tcrossprod(bp_matrix)
diag(login_matrix) <- 0
rm(bp_matrix)

# convert it into a network and removing the loops 
login_network <- graph_from_adjacency_matrix(login_matrix, mode = "undirected", weighted = TRUE)
login_network <- simplify(login_network, remove.loops = TRUE)

# getting the weighted edgelist from 
login_edgelist  <- graph.adjacency(login_matrix, mode = "undirected", weighted=TRUE)
login_edgelist <- get.data.frame(login_edgelist)
head(login_edgelist)
login_edgelist <- login_edgelist %>% 
  filter(weight != 0) %>% 
  rename(Source = from, Target = to, Weight = weight)
login_edgelist %>% arrange(-Weight)



conn <- dbConnect(drv = PostgreSQL(), dbname = "sdad_data", 
                  host = "sdad.policy-analytics.net", port = 5436, 
                  user = Sys.getenv("db_userid"), password = Sys.getenv("db_pwd"))

# query the bipartite edgelist data from github data  
country_codes <- dbGetQuery(conn, "SELECT login, country_code
                                   FROM github.users_gh_cc")

dbDisconnect(conn)

country_codes <- country_codes %>% 
  filter(!grepl(",", country_code)) %>% 
  filter(!grepl("europe", country_code)) %>% 
  filter(!grepl("americas", country_code)) %>% 
  filter(!grepl("asia", country_code)) %>% 
  filter(!grepl("africa", country_code)) %>% 
  mutate(country_code = trimws(country_code))

# changing country codes to country name 
country_codes$country <- countrycode(country_codes$country, origin = 'iso2c', destination = 'country.name')

# changing Kosovo to have a valid country name 
country_codes <- country_codes %>% 
  mutate(country = ifelse(test = str_detect(string = country_code, pattern = "xk"), yes = "Kosovo", no = country)) 

country_edgelist <- login_edgelist %>% 
  rename(login = Source) %>% 
  inner_join(country_codes, by = "login") %>% 
  rename(Source = country) %>% 
  select(Source, Target, Weight)

country_edgelist <- country_edgelist %>% 
  rename(login = Target) %>% 
  inner_join(country_codes, by = "login") %>% 
  rename(Target = country) %>% 
  select(Source, Target)

country_edgelist_08 <- country_edgelist %>% 
  count(Source, Target) %>% 
  arrange(-n)

country_edgelist_08
```

```{r loading edgelist data, warning=FALSE, echo=FALSE}

# connect to postgresql to get data (in rivanna)
conn <- dbConnect(drv = PostgreSQL(), dbname = "sdad_data", 
                  host = "sdad.policy-analytics.net", port = 5436, 
                  user = Sys.getenv("db_userid"), password = Sys.getenv("db_pwd"))

# query the bipartite edgelist data from github data  
bp_edgelist <- dbGetQuery(conn, "SELECT login, slug, year, weight 
                                 FROM github.intl_dyn_bp_el_new 
                                 WHERE year > 2007 AND year < 2014")

# disconnect from postgresql
dbDisconnect(conn)

# this code is taken from network scientists to reduce bipartite networks to single mode networks 
bp_matrix <- spMatrix(nrow=length(unique(bp_edgelist$login)),
              ncol=length(unique(bp_edgelist$slug)),
              i = as.numeric(factor(bp_edgelist$login)),
              j = as.numeric(factor(bp_edgelist$slug)),
              x = rep(1, length(as.numeric(bp_edgelist$login))))
row.names(bp_matrix) <- levels(factor(bp_edgelist$login))
colnames(bp_matrix) <- levels(factor(bp_edgelist$slug))

# we then multiply these incidence matrices and impute zeroes to prevent self loops 
login_matrix <- tcrossprod(bp_matrix)
diag(login_matrix) <- 0
rm(bp_matrix)

# convert it into a network and removing the loops 
login_network <- graph_from_adjacency_matrix(login_matrix, mode = "undirected", weighted = TRUE)
login_network <- simplify(login_network, remove.loops = TRUE)

# getting the weighted edgelist from 
login_edgelist  <- graph.adjacency(login_matrix, mode = "undirected", weighted=TRUE)
login_edgelist <- get.data.frame(login_edgelist)
head(login_edgelist)
login_edgelist <- login_edgelist %>% 
  filter(weight != 0) %>% 
  rename(Source = from, Target = to, Weight = weight)
login_edgelist %>% arrange(-Weight)



conn <- dbConnect(drv = PostgreSQL(), dbname = "sdad_data", 
                  host = "sdad.policy-analytics.net", port = 5436, 
                  user = Sys.getenv("db_userid"), password = Sys.getenv("db_pwd"))

# query the bipartite edgelist data from github data  
country_codes <- dbGetQuery(conn, "SELECT login, country_code
                                   FROM github.users_gh_cc")

dbDisconnect(conn)

country_codes <- country_codes %>% 
  filter(!grepl(",", country_code)) %>% 
  filter(!grepl("europe", country_code)) %>% 
  filter(!grepl("americas", country_code)) %>% 
  filter(!grepl("asia", country_code)) %>% 
  filter(!grepl("africa", country_code)) %>% 
  mutate(country_code = trimws(country_code))

# changing country codes to country name 
country_codes$country <- countrycode(country_codes$country, origin = 'iso2c', destination = 'country.name')

# changing Kosovo to have a valid country name 
country_codes <- country_codes %>% 
  mutate(country = ifelse(test = str_detect(string = country_code, pattern = "xk"), yes = "Kosovo", no = country)) 

country_edgelist <- login_edgelist %>% 
  rename(login = Source) %>% 
  inner_join(country_codes, by = "login") %>% 
  rename(Source = country) %>% 
  select(Source, Target, Weight)

country_edgelist <- country_edgelist %>% 
  rename(login = Target) %>% 
  inner_join(country_codes, by = "login") %>% 
  rename(Target = country) %>% 
  select(Source, Target)

country_edgelist_0813 <- country_edgelist %>% 
  count(Source, Target) %>% 
  arrange(-n)

country_edgelist_0813
```


```{r loading edgelist data, warning=FALSE, echo=FALSE}

# connect to postgresql to get data (in rivanna)
conn <- dbConnect(drv = PostgreSQL(), dbname = "sdad_data", 
                  host = "sdad.policy-analytics.net", port = 5436, 
                  user = Sys.getenv("db_userid"), password = Sys.getenv("db_pwd"))

# query the bipartite edgelist data from github data  
bp_edgelist <- dbGetQuery(conn, "SELECT login, slug, year, weight 
                                 FROM github.intl_dyn_bp_el_new 
                                 WHERE year > 2007 AND year < 2019")

# disconnect from postgresql
dbDisconnect(conn)

# this code is taken from network scientists to reduce bipartite networks to single mode networks 
bp_matrix <- spMatrix(nrow=length(unique(bp_edgelist$login)),
              ncol=length(unique(bp_edgelist$slug)),
              i = as.numeric(factor(bp_edgelist$login)),
              j = as.numeric(factor(bp_edgelist$slug)),
              x = rep(1, length(as.numeric(bp_edgelist$login))))
row.names(bp_matrix) <- levels(factor(bp_edgelist$login))
colnames(bp_matrix) <- levels(factor(bp_edgelist$slug))

# we then multiply these incidence matrices and impute zeroes to prevent self loops 
login_matrix <- tcrossprod(bp_matrix)
diag(login_matrix) <- 0
rm(bp_matrix)

# convert it into a network and removing the loops 
login_network <- graph_from_adjacency_matrix(login_matrix, mode = "undirected", weighted = TRUE)
login_network <- simplify(login_network, remove.loops = TRUE)

# getting the weighted edgelist from 
login_edgelist  <- graph.adjacency(login_matrix, mode = "undirected", weighted=TRUE)
login_edgelist <- get.data.frame(login_edgelist)
head(login_edgelist)
login_edgelist <- login_edgelist %>% 
  filter(weight != 0) %>% 
  rename(Source = from, Target = to, Weight = weight)
login_edgelist %>% arrange(-Weight)



conn <- dbConnect(drv = PostgreSQL(), dbname = "sdad_data", 
                  host = "sdad.policy-analytics.net", port = 5436, 
                  user = Sys.getenv("db_userid"), password = Sys.getenv("db_pwd"))

# query the bipartite edgelist data from github data  
country_codes <- dbGetQuery(conn, "SELECT login, country_code
                                   FROM github.users_gh_cc")

dbDisconnect(conn)

country_codes <- country_codes %>% 
  filter(!grepl(",", country_code)) %>% 
  filter(!grepl("europe", country_code)) %>% 
  filter(!grepl("americas", country_code)) %>% 
  filter(!grepl("asia", country_code)) %>% 
  filter(!grepl("africa", country_code)) %>% 
  mutate(country_code = trimws(country_code))

# changing country codes to country name 
country_codes$country <- countrycode(country_codes$country, origin = 'iso2c', destination = 'country.name')

# changing Kosovo to have a valid country name 
country_codes <- country_codes %>% 
  mutate(country = ifelse(test = str_detect(string = country_code, pattern = "xk"), yes = "Kosovo", no = country)) 

country_edgelist <- login_edgelist %>% 
  rename(login = Source) %>% 
  inner_join(country_codes, by = "login") %>% 
  rename(Source = country) %>% 
  select(Source, Target, Weight)

country_edgelist <- country_edgelist %>% 
  rename(login = Target) %>% 
  inner_join(country_codes, by = "login") %>% 
  rename(Target = country) %>% 
  select(Source, Target)

country_edgelist_0818 <- country_edgelist %>% 
  count(Source, Target) %>% 
  arrange(-n)

country_edgelist_0818
```

```{r}

write_csv(country_edgelist_08, "country_edgelist_08.csv")
write_csv(country_edgelist_0813, "country_edgelist_0813.csv")
write_csv(country_edgelist_0818, "country_edgelist_0818.csv")



```


```{r global network stats}
# loading the network (if the instance crashes while doing network_stats)

setwd("~/oss-2020")
login_edgelist <- read_csv("country_edgelist_08.csv")
login_edgelist <- login_edgelist %>% 
  rename(from = Source,
         to = Target, 
         weight = n)


login_network <- graph.data.frame(login_edgelist, directed = FALSE)


network_stats <- data.frame(year="2008")                                          

# node and edge counts  
network_stats$node_count <- gorder(login_network)  
network_stats$edge_count <- gsize(login_network)
network_stats$commits <- sum(login_edgelist$weight)

# isolates, dyads and triads 
network_stats$isolates <- sum(degree(login_network)==0)
network_stats$dyads <- dyad_census(login_network)[1]
#oss_triads <- triad.census(login_network)
##network_stats$triads_003 <- oss_triads[1] 
#network_stats$triads_102 <- oss_triads[3]
#network_stats$triads_201 <- oss_triads[11]
#network_stats$triads_300 <- oss_triads[16]


# density and transitivity
network_stats$density <- edge_density(login_network, loops=FALSE) 
network_stats$transitivity <- transitivity(login_network)
net_globals <- data.frame(event="net_globals", time=now("EST"))
#time_log <- rbind(time_log, net_globals); rm(net_globals)

# diameter and mean distance (very computationally intensive)
network_stats$diameter <- diameter(login_network, directed = FALSE)
network_stats$mean_distance <- mean_distance(login_network, directed = FALSE)
louvain <- cluster_louvain(login_network)
network_stats$louvain <- modularity(louvain, weights = TRUE)
 
# community detection (using louvain method)
louvain <- cluster_louvain(login_network)
network_stats$louvain <- modularity(louvain)
network_stats$louvain_scaled <- modularity(louvain) / gorder(login_network)
network_stats$louvain_logged <- modularity(louvain) / log(gorder(login_network))
# community detection (using fast & greedy method)
fstgrdy <- fastgreedy.community(login_network)
network_stats$fstgrdy <- modularity(fstgrdy)
network_stats$fstgrdy_scaled <- modularity(fstgrdy) / gorder(login_network)
network_stats$fstgrdy_logged <- modularity(fstgrdy) / log(gorder(login_network))

# decomposition statistics 
decomposition_stats <- table(sapply(decompose.graph(login_network), vcount))

net_comm_det <- data.frame(event="net_comm_det", time=now("EST"))
#time_log <- rbind(time_log, net_comm_det); rm(net_comm_det)

# now, we need to construct a nodelist 
nodelist <- data.frame(id = c(1:(igraph::vcount(login_network))), login = igraph::V(login_network)$name)

# degree, weighted degree, k core and modularity 
node_stats_start <- data.frame(event="node_stats_start", time=now("EST"))
#time_log <- rbind(time_log, node_stats_start); rm(node_stats_start)
nodelist$degree <- degree(login_network)
nodelist$wtd_degree <- strength(login_network)
nodelist$k_core <- coreness(login_network)
#louvain <- cluster_louvain(login_network)
nodelist$louvain_comm <- louvain$membership
#fstgrdy <- fastgreedy.community(login_network)
#nodelist$fstgrdy_comm <- fstgrdy$membership
components <- components(login_network)
nodelist$component <- components$membership


nodelist <- nodelist %>% rename(country = login)

#nodelist <- nodelist %>% rename(country_code = login)
#nodelist$country_code <- toupper(nodelist$country_code)
#nodelist$country <- countrycode(nodelist$country_code, origin = 'iso2c', destination = 'country.name')
#nodelist <- nodelist %>% select(id, country_code, country, everything())
#setwd("~/intl-dyn-ctry")
#write_csv(nodelist, "country_nodelist_all_years.csv")

node_stats_end <- data.frame(event="node_stats_end", time=now("EST"))
#time_log <- rbind(time_log, node_stats_end); rm(node_stats_end, louvain, components)

#setwd("~/intl-dyn-ctry")
#saveRDS(network_stats, "global_netstats_2008-17.rds")  
#saveRDS(decomposition_stats, "decomposition_stats_2008-17.rds") 
#saveRDS(nodelist, "nodelist_2008-17.rds")  
#saveRDS(time_log, "timelog_2008-17.rds")

```
