==> src/05_github-networks/02_international-networks/03_country-networks/00_antiquated/08_ctry-powerlaw-analyses.Rmd <==
---
title: "International Collaboration Netwoks: Fitting Power Laws"
output: html_document
---

```{r loading edgelist data, warning=FALSE, echo=FALSE}
rm(list = ls())
# install.packages(c("tidyverse", "igraph", "visNetwork", "bc3net", 

# load packages 
for (pkg in c("tidyverse", "igraph", "data.table", "R.utils", "RPostgreSQL",
              "cowplot", "maditr", "lubridate", "poweRlaw")) {library(pkg, character.only = TRUE)}

# connect to postgresql to get data (in rivanna)
conn <- dbConnect(drv = PostgreSQL(), 
                  dbname = "sdad", 
                  host = "10.250.124.195", 
                  port = 5432, 
                  user = Sys.getenv("db_userid"), 
                  password = Sys.getenv("db_pwd"))

# query the bipartite edgelist data from github data  
ctry_edgelist <- dbGetQuery(conn, "SELECT country1, country2, repo_wts
                                   FROM gh.sna_intl_ctry_edgelist_0819")

# disconnect from postgresql
dbDisconnect(conn)

# convert edgelist to network 
ctry_edgelist <- ctry_edgelist %>% 
  select(country1, country2, repo_wts) %>% 
  rename(from = country1, to = country2, weight = repo_wts) %>% 
  group_by(from, to) %>% 
  summarize(weight = sum(weight)) %>% 
  arrange(-weight)
  
ctry_network <- simplify(graph.data.frame(ctry_edgelist, directed = FALSE), 
                         remove.loops = FALSE, 
                         edge.attr.comb = igraph_opt("edge.attr.comb"))
is_weighted(ctry_network)
```
```{r fit_powerlaw}
data <- degree(ctry_network)
data <- data[data>0]
data.dist <- data.frame(k=0:max(data),p_k=degree_distribution(ctry_network))
data.dist <- data.dist[data.dist$p_k>0,]
ggplot(data.dist) + geom_point(aes(x=k, y=p_k)) + theme_bw()
```

```{r initial_estimation}
# run initial estimation 
m_pl <- displ$new(data)
est_pl <- estimate_xmin(m_pl)
# initial estimation of kmin
est_pl$xmin  
# initial estimation of γ
est_pl$pars 
# calculate D with Kolgomorov-Smirnov test
est_pl$gof 
```
```{r}
(est = estimate_pars(m_pl))
(est = estimate_xmin(m_pl))
```

```{r}
m_pl$setXmin(est)
## Plot the data (from xmin)
plot(m_pl)
## Add in the fitted distribution
lines(m_pl, col = 2)
dd = plot(m_pl)
head(dd, 5)
```

```{r}
# finish remainder of process using https://rpubs.com/lgadar/power-law
data.s <- unique(data)
d_est <- data.frame(K_min=sort(data.s)[1:(length(data.s)-2)], gamma=rep(0,length(data.s)-2), D=rep(0,length(data.s)-2))
for (i in d_est$K_min){
  d_est[which(d_est$K_min == i),2] <- estimate_xmin(m_pl, xmins = i)$pars
  d_est[which(d_est$K_min == i),3] <- estimate_xmin(m_pl, xmins = i)$gof
}
K.min_D.min <- d_est[which.min(d_est$D), 1]
ggplot(data=d_est, aes(x=K_min, y=D)) + geom_line() + theme_bw() + 
  geom_vline(xintercept=K.min_D.min, colour="red") + annotate("text", x=K.min_D.min, y=max(d_est$D)/3*2, label=K.min_D.min)
ggplot(data=d_est, aes(x=K_min, y=gamma)) + geom_line() + theme_bw() + 
  geom_vline(xintercept=K.min_D.min, colour="red") + annotate("text", x=K.min_D.min, y=max(d_est$gamma)/3*2, label=K.min_D.min)
```

```{r}
m_pl$setXmin(est_pl)
plot.data <- plot(m_pl, draw = F)
fit.data <- lines(m_pl, draw = F)
ggplot(plot.data) + geom_point(aes(x=log(x), y=log(y))) + labs(x="log(k)", y="log(CDF)") + theme_bw() + 
  geom_line(data=fit.data, aes(x=log(x), y=log(y)), colour="red")
```
```{r}
#threads=core number of processor that used by function
#parallel::detectCores() #determines how many cores in your computer
#bs <- bootstrap(m_pl, no_of_sims=1000, threads=15, seed = 123)
#sd(bs$bootstraps[, 2])
##sd(bs$bootstraps[, 3])
#plot(bs, trim = 0.1)
```

```{r}
bs_pl <- bootstrap_p(m_pl, no_of_sims=1000, threads=15, seed = 123)
df_bs_pl <- bs_pl$bootstraps
ggplot(data=df_bs_pl, aes(pars)) + geom_histogram() + labs(x="alpha", y="frequency") + theme_bw()
ggplot(data=df_bs_pl, aes(xmin)) + geom_histogram() + labs(x="K_min", y="frequency") + theme_bw()
```
```{r}
gamma_D.min <- d_est[which.min(d_est$D), 2]
ggplot(data=df_bs_pl, aes(x=xmin, y=pars)) + labs(x="K_min", y="alpha") + theme_bw() + 
  geom_point(shape=21, colour="black", fill="red", size=0.5, stroke=2, 
             position = position_jitter(), alpha=0.6) +
  geom_vline(xintercept=K.min_D.min, colour="blue") +
  geom_hline(yintercept=gamma_D.min, colour="blue") +
  annotate("text", x=K.min_D.min, y=min(df_bs_pl$pars), label=K.min_D.min, col="blue") +
  annotate("text", x=min(df_bs_pl$xmin), y=gamma_D.min, label=round(gamma_D.min, digits=2), col="blue")

D.min <- d_est[which.min(d_est$D), 3]
ggplot(data=df_bs_pl, aes(gof)) + geom_histogram() + labs(x="D", y="frequency") + geom_vline(xintercept=D.min, colour="red") + theme_bw()
```

```{r power_law_best_fit}
bs_pl$p
bs_pl$gof
```
To determine when we would reject or fail to reject for each KS test, we will use the significance level of 0.10, which means if the p value is greater than .10, we fail to reject the null and conclude that both data sets come from the same distribution, whereas if the p-value is less than or equal to 0.10, we reject the null and conclude that the data sets do not come from a power law distribution. 

```{r}
#lognormal
m_ln = dislnorm$new(data)
est_ln <- estimate_xmin(m_ln)
m_ln$setXmin(est_ln)

#exponential
m_exp = disexp$new(data)
est_exp <- estimate_xmin(m_exp)
m_exp$setXmin(est_exp)

#poisson
m_poi = dispois$new(data)
est_poi <- estimate_xmin(m_poi)
m_poi$setXmin(est_poi)

plot(m_pl)
lines(m_pl, col="red")
lines(m_ln, col="green")
lines(m_poi, col="blue")
lines(m_exp, col="magenta")
```

```{r}
#bs_ln = bootstrap(m_ln)
#m_ln$setXmin(m_pl$getXmin())
#est_ln = estimate_pars(m_ln)
#m_ln$setPars(est_ln)
#comp = compare_distributions(m_pl, m_ln)
#comp$p_one_sided
#comp$p_two_sided
```

```{r}
year <- "2008-19"
alpha <- est$pars
alpha_sd <- sd(bs_pl$bootstraps$pars)
xmin <- est$xmin
xmin_sd <- sd(bs_pl$bootstraps$xmin)
gof <- bs_pl$gof
pvalue <- bs_pl$p
power_law_data <- as.data.frame(cbind(year, alpha, alpha_sd, xmin, xmin_sd, pvalue, gof))
setwd("~/oss-data/intl-ctry-nets-cum/02_power-law-data")
saveRDS(power_law_data, str_c("power_law_data_",year,".rds"))
```

# Aggregate Results 

```{r}
setwd("~/oss-data/intl-ctry-nets-cum/02_power-law-data")
powerlaw_2008 <- read_rds("power_law_data_2008.rds")
powerlaw_2009 <- read_rds("power_law_data_2008-09.rds")
powerlaw_2010 <- read_rds("power_law_data_2008-10.rds")
powerlaw_2011 <- read_rds("power_law_data_2008-11.rds")
powerlaw_2012 <- read_rds("power_law_data_2008-12.rds")
powerlaw_2013 <- read_rds("power_law_data_2008-13.rds")
powerlaw_2014 <- read_rds("power_law_data_2008-14.rds")
powerlaw_2015 <- read_rds("power_law_data_2008-15.rds")
powerlaw_2016 <- read_rds("power_law_data_2008-16.rds")
powerlaw_2017 <- read_rds("power_law_data_2008-17.rds")
powerlaw_2018 <- read_rds("power_law_data_2008-18.rds")
powerlaw_2019 <- read_rds("power_law_data_2008-19.rds")
power_law_data_all <- rbind(powerlaw_2008, powerlaw_2009, powerlaw_2010, powerlaw_2011, powerlaw_2012, powerlaw_2013,
                            powerlaw_2014, powerlaw_2015, powerlaw_2016, powerlaw_2017, powerlaw_2018, powerlaw_2019)
setwd("~/oss-data/intl-ctry-nets-cum/02_power-law-data")
saveRDS(power_law_data_all, "power_law_data_all.rds")
```

## References 

Blondel, V. D., Guillaume, J. L., Lambiotte, R., & Lefebvre, E. (2008). Fast unfolding of communities in large networks. Journal of statistical mechanics: theory and experiment, 2008(10), P10008.

http://pablobarbera.com/big-data-upf/html/02b-networks-descriptive-analysis.html
https://igraph.org/r/doc/cluster_louvain.html
https://kateto.net/networks-r-igraph 
https://sites.fas.harvard.edu/~airoldi/pub/books/BookDraft-CsardiNepuszAiroldi2016.pdf
https://rpubs.com/lgadar/power-law 
http://recon.icm.edu.pl/wp-content/uploads/2019/05/linkprediction.pdf
https://cran.r-project.org/web/packages/poweRlaw/vignettes/a_introduction.pdf
https://cran.r-project.org/web/packages/poweRlaw/vignettes/b_powerlaw_examples.pdf
https://epubs.siam.org/doi/pdf/10.1137/070710111?casa_token=XROmDjUk7MsAAAAA%3AHTfw7zmR_dZO0ef1kETfiD27rsTeqbh88DZ2hRV7tv5lGtB-lgny7rOlxi_7Htv3A30gXpgxlnrl&
